+++ app/components/files/select-files.js
@@ -25,13 +25,13 @@ export default class SelectFiles extends SafeComponent {
             flexGrow: 1,
             flexDirection: 'row',
             alignItems: 'center',
-            padding: 4,
+            padding: vars.spacing.small.mini2x,
             paddingTop: 0,
             paddingBottom: 0
         };
         const style = {
             flexGrow: 1,
-            marginLeft: 8
+            marginLeft: vars.spacing.small.midi2x
         };
 
         return (
@@ -53,7 +53,7 @@ export default class SelectFiles extends SafeComponent {
             flexGrow: 1,
             flexDirection: 'row',
             alignItems: 'center',
-            padding: 4,
+            padding: vars.spacing.small.mini2x,
             paddingTop: vars.statusBarHeight,
             paddingBottom: 0
         };
@@ -61,12 +61,12 @@ export default class SelectFiles extends SafeComponent {
             flexGrow: 1
         };
         const textStyle = {
-            fontSize: 14,
+            fontSize: vars.font.size.normal,
             fontWeight: vars.font.weight.semiBold,
             color: 'rgba(0, 0, 0, .54)'
         };
         const goStyle = {
-            fontSize: 14,
+            fontSize: vars.font.size.normal,
             fontWeight: vars.font.weight.semiBold,
             color: vars.bg
         };