+++ app/components/messaging/chat-section-header.js
@@ -12,8 +12,8 @@ export default class ChatSectionHeader extends SafeComponent {
     renderThrow() {
         const { title, collapsible } = this.props;
         const style = {
-            paddingLeft: 18,
-            paddingRight: 16,
+            paddingLeft: vars.spacing.medium.midi,
+            paddingRight: vars.spacing.medium.mini2x,
             height: 48,
             justifyContent: 'space-between',
             backgroundColor: vars.lightGrayBg,