+++ app/components/settings/settings-level-2.js
@@ -105,8 +105,8 @@ export default class SettingsLevel2 extends SafeComponent {
     preferences = () => {
         const text = {
             color: vars.txtMedium,
-            marginBottom: 8,
-            marginLeft: 8
+            marginBottom: vars.spacing.small.midi2x,
+            marginLeft: vars.spacing.small.midi2x
         };
 
         return (